Rename kebab-case files to snake_case
