+++ rollup.config.js
@@ -42,7 +42,7 @@ export default [
 				generateBundle(options, bundle) {
 					const mod = bundle['internal.mjs'];
 					if (mod) {
-						fs.writeFileSync('src/compile/internal-exports.ts', `// This file is automatically generated\nexport default new Set(${JSON.stringify(mod.exports)});`);
+						fs.writeFileSync('src/compile/internal_exports.ts', `// This file is automatically generated\nexport default new Set(${JSON.stringify(mod.exports)});`);
 					}
 				}
 			}]
